光影設定 3/17
--新增--
新增音效--左輪射擊&換彈 命中怪物 拾取武器
光影設定--部分光源在靠近時才會啟動光影與調整品質

--修正--
怪物彈孔--修改命中怪物彈孔
螃蟹材質--修正螃蟹材質

+++ Assets/AA/Scripts/system/Ammunition.cs
@@ -24,7 +24,7 @@ public class Ammunition : MonoBehaviour
         Color.a = 0;
         RawImages[0].gameObject.SetActive(false);
         RawImages[1].gameObject.SetActive(false);
-        animator.SetFloat("WeapSW", -1);
+        animator.SetInteger("WeapSW", -1);
 
     }
